split Uniqify module
