+++ liquid-fixpoint.cabal
@@ -172,7 +172,8 @@ Library
                    Language.Fixpoint.Solver.Monad,
                    Language.Fixpoint.Solver.Deps,
                    Language.Fixpoint.Solver.TrivialSort,
-                   Language.Fixpoint.Solver.Uniqify,
+                   Language.Fixpoint.Solver.UniqifyKVars,
+                   Language.Fixpoint.Solver.UniqifyBinds,
                    Language.Fixpoint.Solver.Eliminate,
                    Language.Fixpoint.Solver.Validate,
                    Language.Fixpoint.Utils.Statistics,
