+++ src/Language/Fixpoint/Solver/Eliminate.hs
@@ -5,9 +5,9 @@ module Language.Fixpoint.Solver.Eliminate
        (eliminateAll) where
 
 import           Language.Fixpoint.Types
-import           Language.Fixpoint.Types.Names           (existSymbol)
+import           Language.Fixpoint.Types.Names     (existSymbol)
+import           Language.Fixpoint.Types.Visitor   (kvars)
 import           Language.Fixpoint.Solver.Deps     (depNonCuts, deps)
-import           Language.Fixpoint.Types.Visitor         (kvars)
 import           Language.Fixpoint.Misc            (fst3)
 import           Language.Fixpoint.Solver.Solution (Solution, mkJVar)
 
@@ -39,7 +39,7 @@ extractPred kDom be sc = renameQuantified (subcId sc) kSol
   where
     env = clhs be sc
     binds = second sr_sort <$> env
-    nonFuncBinds = filter (fst . (first $ nonFunction be)) binds
+    nonFuncBinds = filter (nonFunction be . fst) binds
     lhsPreds = bindPred <$> env
     suPreds = substPreds kDom $ crhs sc
     kSol = PExist nonFuncBinds $ PAnd (lhsPreds ++ suPreds)
