+++ src/Language/Fixpoint/Solver/UniqifyBinds.hs
@@ -1,33 +1,32 @@
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE PatternGuards #-}
 
-module Language.Fixpoint.Solver.Uniqify (renameAll) where
+-- This module makes it so no binds with different sorts have the same name.
+
+module Language.Fixpoint.Solver.UniqifyBinds (renameAll) where
 
 import           Language.Fixpoint.Types
-import           Language.Fixpoint.Solver.Types     (CId)
-import           Language.Fixpoint.Types.Visitor    (mapKVarSubsts)
-import           Language.Fixpoint.Types.Names      (renameSymbol, kArgSymbol)
-import           Language.Fixpoint.Misc       (fst3, mlookup)
-import qualified Data.HashMap.Strict                as M
-import qualified Data.HashSet                       as S
-import qualified Data.List                          as L
-import           Data.Foldable                      (foldl')
-import           Data.Maybe                         (catMaybes, fromJust, isJust)
-import           Data.Hashable                      (Hashable)
-import           GHC.Generics                       (Generic)
-import           Control.Arrow                      (second)
-import           Control.DeepSeq                    (NFData, ($!!))
+import           Language.Fixpoint.Solver.Types  (CId)
+import           Language.Fixpoint.Misc          (fst3, mlookup)
+
+import qualified Data.HashMap.Strict as M
+import qualified Data.HashSet        as S
+import qualified Data.List           as L
+import           Data.Foldable       (foldl')
+import           Data.Maybe          (catMaybes, fromJust, isJust)
+import           Data.Hashable       (Hashable)
+import           GHC.Generics        (Generic)
+import           Control.Arrow       (second)
+import           Control.DeepSeq     (NFData, ($!!))
 
 --------------------------------------------------------------
 renameAll    :: SInfo a -> SInfo a
-renameAll fi0 = fi4
+renameAll fi2 = fi4
   where
     fi4      = {-# SCC "renameBinds" #-} renameBinds fi3 $!! rnm
     fi3      = {-# SCC "renameVars"  #-} renameVars fi2 rnm $!! idm
     rnm      = {-# SCC "mkRenameMap" #-} mkRenameMap $!! bs fi2
     idm      = {-# SCC "mkIdMap"     #-} mkIdMap fi2
-    fi2      = {-# SCC "updateWfcs"  #-} updateWfcs fi1
-    fi1      = {-# SCC "remakeSubsts" #-} remakeSubsts fi0
 --------------------------------------------------------------
 
 data Ref = RB BindId | RI CId deriving (Eq, Generic)
@@ -135,71 +134,3 @@ renameBind m (i, sym, sr)
     t       = sr_sort sr
     mnewSym = fromJust $ L.lookup t $ mlookup m sym
 --------------------------------------------------------------
-
---------------------------------------------------------------
-remakeSubsts :: SInfo a -> SInfo a
---------------------------------------------------------------
-remakeSubsts fi = mapKVarSubsts (remakeSubstIfWfcExists fi) fi
-
---TODO: why are there kvars with no WfC?
-remakeSubstIfWfcExists :: SInfo a -> KVar -> Subst -> Subst
-remakeSubstIfWfcExists fi k su
-  | k `M.member` ws fi = remakeSubst fi k su
-  | otherwise          = Su M.empty
-
-remakeSubst :: SInfo a -> KVar -> Subst -> Subst
-remakeSubst fi k su = foldl' (updateSubst k) su kDom
-  where
-    w    = (ws fi) M.! k
-    kDom = (fst3 $ wrft w) : (fst <$> envCs (bs fi) (wenv w))
-
-updateSubst :: KVar -> Subst -> Symbol -> Subst
-updateSubst k (Su su) sym
-  | sym `M.member` su = Su $ M.delete sym $ M.insert (kArgSymbol' sym k) (su M.! sym) su
-  | otherwise         = Su $ M.insert (kArgSymbol' sym k) (eVar sym) su
---------------------------------------------------------------
-
---------------------------------------------------------------
-updateWfcs :: SInfo a -> SInfo a
---------------------------------------------------------------
-updateWfcs fi = M.foldl' (updateWfc $ bs fi) fi (ws fi)
-
-updateWfc :: BindEnv -> SInfo a -> WfC a -> SInfo a
-updateWfc be fi w = fi' { ws = M.insert k w' (ws fi) }
-  where
-    (v, t, k) = wrft w
-    (fi', newIds) = insertNewBinds w fi k
-    env' = insertsIBindEnv newIds emptyIBindEnv
-    w' = w { wenv = env', wrft = (kArgSymbol' v k, t, k) }
-
-insertNewBinds :: WfC a -> SInfo a -> KVar -> (SInfo a, [BindId])
-insertNewBinds w fi k = foldl' (accumBindsIfValid k) (fi, []) (elemsIBindEnv $ wenv w)
-
-accumBindsIfValid :: KVar -> (SInfo a, [BindId]) -> BindId -> (SInfo a, [BindId])
-accumBindsIfValid k (fi, ids) i = if renamable then accumBinds k (fi, ids) i else (fi, i : ids)
-  where
-    (oldSym, sr) = lookupBindEnv i (bs fi)
-    renamable = isValidInRefinements $ sr_sort sr
-
-accumBinds :: KVar -> (SInfo a, [BindId]) -> BindId -> (SInfo a, [BindId])
-accumBinds k (fi, ids) i = (fi {bs = be'}, i' : ids)
-  where
-    --TODO: could we ignore the old SortedReft? what would it mean if it were non-trivial in a wf environment?
-    (oldSym, sr) = lookupBindEnv i (bs fi)
-    newSym = kArgSymbol' oldSym k
-    (i', be') = insertBindEnv newSym sr (bs fi)
---------------------------------------------------------------
-
-kArgSymbol' :: Symbol -> KVar -> Symbol
-kArgSymbol' sym k = (kArgSymbol sym) `mappend` (symbol "_") `mappend` (kv k)
-
-isValidInRefinements :: Sort -> Bool
-isValidInRefinements FInt        = True
-isValidInRefinements FReal       = True
-isValidInRefinements FNum        = False
-isValidInRefinements FFrac       = False
-isValidInRefinements (FObj _)    = True
-isValidInRefinements (FVar _)    = True
-isValidInRefinements (FFunc _ _) = False
-isValidInRefinements (FTC _)     = True --TODO is this true? seems to be required for e.g. ResolvePred.hs
-isValidInRefinements (FApp _ _)  = True
